refactor: extract shared helpers, fix naming, split god file
Structural refactoring verified safe with Regrets regression testing.

## Changes

### Decomposition — Split god file
- Split `sortingAlgorithms.js` (4 algorithms in one file) into individual files:
  - `bubble-sort.js` — Bubble sort algorithm
  - `selection-sort.js` — Selection sort algorithm
  - `insertion-sort.js` — Insertion sort algorithm
- `sortingAlgorithms.js` now re-exports from split modules for backward compatibility

### Shared Helpers — Eliminate duplication
- Created `helpers/graph-helpers.js` with `buildPathActions()` function
- Previously, identical `pathActions()` was duplicated in both `graph.js` and `shortestPath.js`
- Both files now import from shared helper, export as `pathActions` for backward compat
- Fixed ESM import: `shortestPath.js` changed `'./graph'` → `'./graph.js'`

### Naming — Make names honest and descriptive
- Fixed typo: `seive()` → `sieve()` (Sieve of Eratosthenes) with backward-compatible alias
- Fixed naming convention: `convex_hull()` → `convexHull()` (camelCase in JS) with alias
- Descriptive names: `cw()` → `isClockwiseTurn()`, `ccw()` → `isCounterClockwiseTurn()` with aliases

## Verification

All changes verified safe with Regrets regression testing:
- 15 clusters: all GREEN
- 27 raw outputs: match pre-refactor KEBENARAN 1 exactly
- 15 fingerprints: match pre-refactor KEBENARAN 2
- 2 chain hashes: match pre-refactor chains
